Move index pages
